Restructuring top-level heading names.
I decided "Getting Started" was not the right name for that section. "Getting Started" should be a section with guided examples, not a general overview/reference like this. In view of that, I have renamed "Introduction" to "Welcome", and "Getting Started" to "Overview". I plan to create another section called "Getting Started" or "Quick Start" when the time comes to write some guided tutorials, or shoot some guided tutorial videos.