Exclude developer reference material from search results
I just noticed my translated version of ra3xdh's tutorial is appearing in search results on the HTML site, despite not being in the table of contents for Sphinx. I didn't mean for this to happen, so I'm excluding it forcefully from Sphinx (to avoid confusing people on the docs site, since it's not accessible from the normal table of contents).